requirements.txt
flask
pydub
pyannote.audio
torch
torchaudio
